feat(cb2-4163): refactor
