+++ src/app/forms/templates/hgv/hgv-gross-train-weights.template.ts
@@ -1,5 +1,5 @@
 import { FormNode, FormNodeTypes, FormNodeViewTypes } from '../../services/dynamic-form.types';
-import { generateWeights } from '../general/hgv-weights.template';
+import { generateWeights } from '../general/weights.template';
 
 export const HgvGrossTrainWeight: FormNode = {
   name: 'grossTrainWeight',
@@ -7,5 +7,5 @@ export const HgvGrossTrainWeight: FormNode = {
   value: '',
   type: FormNodeTypes.GROUP,
   viewType: FormNodeViewTypes.SUBHEADING,
-  children: generateWeights(true, 'train')
+  children: generateWeights(true, 'hgv', 'train')
 };
