+++ src/app/forms/templates/hgv/hgv-gross-vehicle-weights.template.ts
@@ -1,5 +1,5 @@
 import { FormNode, FormNodeTypes, FormNodeViewTypes } from '../../services/dynamic-form.types';
-import { generateWeights } from '../general/hgv-weights.template';
+import { generateWeights } from '../general/weights.template';
 
 export const HgvGrossVehicleWeight: FormNode = {
   name: 'grossVehicleWeight',
@@ -7,5 +7,5 @@ export const HgvGrossVehicleWeight: FormNode = {
   value: '',
   type: FormNodeTypes.GROUP,
   viewType: FormNodeViewTypes.SUBHEADING,
-  children: generateWeights(true, 'gross')
+  children: generateWeights(true, 'hgv', 'gross')
 };
