+++ src/app/forms/templates/hgv/hgv-max-train-weights.template.ts
@@ -1,5 +1,5 @@
 import { FormNode, FormNodeTypes, FormNodeViewTypes } from '../../services/dynamic-form.types';
-import { generateWeights } from '../general/hgv-weights.template';
+import { generateWeights } from '../general/weights.template';
 
 export const HgvMaxTrainWeight: FormNode = {
   name: 'maxTrainWeight',
@@ -7,5 +7,5 @@ export const HgvMaxTrainWeight: FormNode = {
   value: '',
   type: FormNodeTypes.GROUP,
   viewType: FormNodeViewTypes.SUBHEADING,
-  children: generateWeights(true, 'max')
+  children: generateWeights(true, 'hgv', 'max')
 };
